requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
pyyaml>=6.0
requests>=2.28.0
